added methods for returning the players ranking
created a method that returns the percentage of all the players on the
system
created a method that returns the percentage and name of the player with
lower percentage
created a method that returns the percentage and name of the player with
higher percentage

also:
changed the mean results to double

+++ src/main/java/com/daus/models/Game.java
@@ -52,7 +52,7 @@ public class Game {
 	public void setDie2Result(int die2Result) {
 		this.die2Result = die2Result;
 	}
-	public int getTotalResult() {
+	public double getTotalResult() {
 		return totalResult;
 	}
 	public void setTotalResult(int totalResult) {
